backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
httpx==0.27.0
beautifulsoup4==4.12.3
requests==2.32.3
python-dotenv==1.0.1
supabase==2.5.0
groq==0.9.0
pydantic==2.7.1
newspaper3k==0.2.8
lxml==5.2.2
lxml_html_clean==0.1.1
python-multipart==0.0.9
slowapi==0.1.9
